requirements.txt
APScheduler
flask
flask_paginate
fpdf
pandas
python-dotenv
Werkzeug
gunicorn